Update auth feature to match new pattern
