+++ src/components/Header/Header.module.scss
@@ -30,7 +30,7 @@
 }
 
 .nav-link {
-  @apply block px-4 py-2 capitalize rounded mt-1;
+  @apply block px-4 py-2 capitalize rounded mt-1 w-full text-left;
 
   color: $secondary;
