--- src/features/auth/components/SignIn/SignIn.js
@@ -1,22 +1,0 @@
-import { Link } from 'react-router-dom';
-
-import Form from './SignInForm';
-
-import styles from '../SignUp/SignUp.module.scss';
-
-const SignIn = () => (
-  <div className="w-full h-100">
-    <h1 className="text-4xl md:text-2xl font-bold leading-tight mt-12 text-center uppercase">
-      Iniciar sesión
-    </h1>
-    <Form />
-    <p className="mt-8 text-center">
-      ¿Aun no tienes una cuenta?{' '}
-      <Link to="/registro" className={styles.link}>
-        Registrarte
-      </Link>
-    </p>
-  </div>
-);
-
-export { SignIn };
