--- src/features/auth/components/SignIn/index.js
@@ -1,1 +1,0 @@
-export { SignIn } from './SignIn';
