+++ src/features/auth/components/SignInForm/SignInForm.js
@@ -1,15 +1,12 @@
-import { useCallback, useMemo } from 'react';
+import { useMemo } from 'react';
 import { useForm } from 'react-hook-form';
 import { string, object } from 'yup';
 import { yupResolver } from '@hookform/resolvers/yup';
+import PropTypes from 'prop-types';
 
 import { Form } from 'features/common';
-import { handleErrors } from 'helpers/errors';
-import { useSignIn } from 'hooks/mutations/auth';
 
-const SignInForm = () => {
-  const { mutate, isLoading } = useSignIn();
-
+const SignInForm = ({ onSubmit, isLoading }) => {
   const schema = useMemo(
     () =>
       object().shape({
@@ -21,16 +18,10 @@ const SignInForm = () => {
 
   const methods = useForm({ resolver: yupResolver(schema) });
 
-  const onSubmit = useCallback(
-    async (values) =>
-      mutate(values, {
-        onError: (errors) => handleErrors(errors, methods.setError),
-      }),
-    [mutate, methods.setError]
-  );
+  const onSubmitHandler = onSubmit(methods.setError);
 
   return (
-    <Form onSubmit={onSubmit} methods={methods}>
+    <Form onSubmit={onSubmitHandler} methods={methods}>
       <Form.Input
         name="email"
         label="Email"
@@ -51,4 +42,9 @@ const SignInForm = () => {
   );
 };
 
-export default SignInForm;
+SignInForm.propTypes = {
+  onSubmit: PropTypes.func.isRequired,
+  isLoading: PropTypes.bool.isRequired,
+};
+
+export { SignInForm };
