+++ src/features/auth/components/SignInForm/index.js
@@ -1,0 +1,1 @@
+export { SignInForm } from './SignInForm';
