--- src/features/auth/components/SignUp/index.js
@@ -1,1 +1,0 @@
-export { SignUp } from './SignUp';
