+++ src/features/auth/components/SignUpForm/index.js
@@ -1,0 +1,1 @@
+export { SignUpForm } from './SignUpForm';
