--- src/features/auth/components/UpdateAccount/index.js
@@ -1,1 +1,0 @@
-export { UpdateAccount } from './UpdateAccount';
