+++ src/features/auth/components/UpdateAccountForm/index.js
@@ -1,0 +1,1 @@
+export { UpdateAccountForm } from './UpdateAccountForm';
