+++ src/features/auth/components/index.js
@@ -1,6 +1,6 @@
 export * from './AccountUserInfo';
-export * from './SignIn';
-export * from './SignUp';
-export * from './UpdateAccount';
+export * from './SignInForm';
+export * from './SignUpForm';
+export * from './UpdateAccountForm';
 export * from './UserDonations';
 export * from './UserPetitions';
