+++ src/features/auth/hooks/index.js
@@ -1,1 +1,2 @@
 export * from './account';
+export * from './auth';
