+++ src/pages/SignIn/SignInPage.module.scss
@@ -1,4 +1,4 @@
-@import '../../../../sass/variables';
+@import '../../sass/variables';
 
 .link {
   @apply hover:opacity-90 font-semibold underline;
